Put old filed back
